+++ publish-service/src/main/java/com/ericsson/eiffel/remrem/publish/controller/ProducerController.java
@@ -1,5 +1,5 @@
 /*
-    Copyright 2017 Ericsson AB.
+    Copyright 2018 Ericsson AB.
     For a full list of individual contributors, please see the commit history.
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -16,6 +16,10 @@ package com.ericsson.eiffel.remrem.publish.controller;
 
 import java.util.Map;
 
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -47,6 +51,7 @@ import ch.qos.logback.classic.Logger;
 
 @RestController
 @RequestMapping("/*")
+@Api(value = "REMReM Publish", description = "REST API for publishing Eiffel messages to message bus")
 public class ProducerController {
 
     @Value("${generate.server.host}")
@@ -78,12 +83,18 @@ public class ProducerController {
     Logger log = (Logger) LoggerFactory.getLogger(ProducerController.class);
 
     @SuppressWarnings({ "rawtypes", "unchecked" })
+    @ApiOperation(value = "To publish eiffel event to message bus", response = String.class)
+    @ApiResponses(value = { @ApiResponse(code = 200, message = "Event sent successfully"),
+            @ApiResponse(code = 400, message = "Invalid event content"),
+            @ApiResponse(code = 404, message = "RabbitMq properties not found"),
+            @ApiResponse(code = 500, message = "Internal server error"),
+            @ApiResponse(code = 503, message = "Service Unavailable") })
     @RequestMapping(value = "/producer/msg", method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity send(@RequestParam(value = "mp", required = false) String msgProtocol,
-            @RequestParam(value = "ud", required = false) String userDomain,
+    public ResponseEntity send(@RequestParam(value = "msgProtocol", required = false) String msgProtocol,
+            @RequestParam(value = "userDomain", required = false) String userDomain,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "rk", required = false) String routingKey, @RequestBody JsonElement body) {
+            @RequestParam(value = "routingKey", required = false) String routingKey, @RequestBody JsonElement body) {
         MsgService msgService = PublishUtils.getMessageService(msgProtocol, msgServices);
 
         log.debug("mp: " + msgProtocol);
@@ -100,35 +111,41 @@ public class ProducerController {
      * This controller provides single RemRem REST API End Point for both RemRem
      * Generate and Publish.
      * 
-     * @param mp
+     * @param msgProtocol
      *            message protocol (required)
      * @param msgType
      *            message type (required)
-     * @param ud
+     * @param userDomain
      *            user domain (not required)
      * @param tag
      *            (not required)
-     * @param rk
+     * @param routingKey
      *            (not required)
      * @return A response entity which contains http status and result
-     * @exception IOException
-     *                On input error.
-     * @see IOException
+//     * @exception IOException
+//     *                On input error.
+//     * @see IOException
      * @use A typical CURL command: curl -H "Content-Type: application/json" -X POST
      *      --data "@inputGenerate_activity_finished.txt"
      *      "http://localhost:8986/generateAndPublish/?mp=eiffelsemantics&msgType=EiffelActivityFinished"
      */
 
     @SuppressWarnings({ "rawtypes", "unchecked" })
+    @ApiOperation(value = "To generate and publish eiffel event to message bus", response = String.class)
+    @ApiResponses(value = { @ApiResponse(code = 200, message = "Event sent successfully"),
+            @ApiResponse(code = 400, message = "Invalid event content"),
+            @ApiResponse(code = 404, message = "RabbitMq properties not found"),
+            @ApiResponse(code = 500, message = "Internal server error"),
+            @ApiResponse(code = 503, message = "Message protocol is invalid") })
     @RequestMapping(value = "/generateAndPublish", method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity generateAndPublish(@RequestParam(value = "mp") String mp, @RequestParam("msgType") String msgType,
-            @RequestParam(value = "ud", required = false) String userDomain,
+    public ResponseEntity generateAndPublish(@RequestParam(value = "msgProtocol") String msgProtocol, @RequestParam("msgType") String msgType,
+            @RequestParam(value = "userDomain", required = false) String userDomain,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "rk", required = false) String routingKey, @RequestBody JsonObject bodyJson) {
+            @RequestParam(value = "routingKey", required = false) String routingKey, @RequestBody JsonObject bodyJson) {
 
         URLTemplate urlTemplate = new URLTemplate();
-        urlTemplate.generate(mp, msgType, userDomain, routingKey, tag, generateServerHost ,generateServerPort);
+        urlTemplate.generate(msgProtocol, msgType, userDomain, routingKey, tag, generateServerHost ,generateServerPort);
 
         ResponseEntity<String> response = null;
 
@@ -147,13 +164,13 @@ public class ProducerController {
             String responseBody = response.getBody();
             // publishing requires an array if you want status code
             responseBody = "[" + responseBody + "]";
-            MsgService msgService = PublishUtils.getMessageService(mp, msgServices);
+            MsgService msgService = PublishUtils.getMessageService(msgProtocol, msgServices);
 
-            log.debug("mp: " + mp);
+            log.debug("msgProtocol: " + msgProtocol);
             log.debug("body: " + responseBody);
             log.debug("user domain suffix: " + userDomain + " tag: " + tag + " Routing Key: " + routingKey);
-            if (msgService != null && mp != null) {
-                rmqHelper.rabbitMqPropertiesInit(mp);
+            if (msgService != null && msgProtocol != null) {
+                rmqHelper.rabbitMqPropertiesInit(msgProtocol);
             }
             SendResult result = messageService.send(responseBody, msgService, userDomain, tag, routingKey);
             return new ResponseEntity(result, messageService.getHttpStatus());
@@ -168,8 +185,7 @@ public class ProducerController {
      * @return this method returns the current version of publish and all loaded
      *         protocols.
      */
-
-
+    @ApiOperation(value = "To get versions of publish and all loaded protocols.", response = String.class)
     @RequestMapping(value = "/versions", method = RequestMethod.GET)
     public JsonElement getVersions() {
         JsonParser parser = new JsonParser();